Change semantics of barkeep ed to be more useful.
This introduces the ability for barkeep ed to make multiple edits in a
single innvocation although still limited to one entry at a time.  A new
issue caused by this is that the entry is edited unconditionally even if
all a user wanted to do was update the cover image and I don't want to
add a -n no edit option just for this.

Other than that, lots of little improvements in quoting and logic.