taverner: support user defined menus
This feature adds the ability for users to provide a printf-like
formatter string using %t and %c placeholders for the entry title and
cover respectively.

Literal % and \ characters are escaped according to printf's rules and
the result is passed directly to printf.

Finally this includes support for user provided commands using the new
-c and -a flags.